[MID-6] changed files structure, made filters displayed by default
