--- src/app/_components/AdminPage/Filters/index.js
@@ -1,1 +1,0 @@
-export * from './Filters';
