--- src/app/_components/AdminPage/OrganizationShow/index.js
@@ -1,1 +1,0 @@
-export { OrganizationShow } from './OrganizationShow';
