--- src/app/_components/AdminPage/OrganizationsList/OrganizationsList.js
@@ -1,18 +1,0 @@
-import React from 'react';
-import {
-  Datagrid, List, TextField, BooleanField, DateField,
-} from 'react-admin';
-import { organizationFilters } from '../Filters';
-
-export function OrganizationsList() {
-  return (
-    <List sort={{ field: 'createdAt', order: 'DESC' }} filters={organizationFilters}>
-      <Datagrid rowClick="show">
-        <TextField source="name" />
-        <DateField showTime source="createdAt" />
-        <TextField source="formatOfWork" />
-        <BooleanField source="isActive" />
-      </Datagrid>
-    </List>
-  );
-}
