--- src/app/_components/AdminPage/OrganizationsList/index.js
@@ -1,1 +1,0 @@
-export { OrganizationsList } from './OrganizationsList';
