--- src/app/_components/AdminPage/SpecialistShow/index.js
@@ -1,1 +1,0 @@
-export { SpecialistShow } from './SpecialistShow';
