--- src/app/_components/AdminPage/SpecialistsList/index.js
@@ -1,1 +1,0 @@
-export { SpecialistsList } from './SpecialistList';
