+++ src/app/admin/_components/organization/index.js
@@ -1,0 +1,2 @@
+export { OrganizationsList } from './List';
+export { OrganizationShow } from './Show';
