+++ src/app/admin/_components/specialist/index.js
@@ -1,0 +1,2 @@
+export { SpecialistsList } from './List';
+export { SpecialistShow } from './Show';
