+++ src/app/api/admin/[...slug]/route.js
@@ -4,11 +4,16 @@ import { prisma } from '@/lib/db';
 import { auth } from '@/lib/auth';
 import { withErrorHandler } from '@/lib/errors/errorHandler';
 import { NotAuthorizedException } from '@/lib/errors/NotAuthorizedException';
-import { MODEL_INCLUDES } from '@/lib/consts';
-import { searchInputFilters } from '@/lib/searchInputFilters';
+import { MODEL_INCLUDES, MODEL_SEARCH_FIELDS } from '@/lib/consts';
+
+export function searchInputFilters(modelName, filter) {
+  if (!filter) return {};
+  const filters = MODEL_SEARCH_FIELDS[modelName].map(field => ({ [field]: { contains: filter, mode: 'insensitive' } }));
+  return { OR: filters };
+}
 
 const handler = auth(
-  withErrorHandler(async (req) => {
+  withErrorHandler(async req => {
     if (!req.auth) throw new NotAuthorizedException();
     const json = await req.json();
     const { resource: modelName } = json;
